requirements.txt
django==5.0.7
djangorestframework
djangorestframework-simplejwt
requests
pyyaml
django-cors-headers
 